raftstore: extract region_info_accessor
