+++ src/raftstore/coprocessor/mod.rs
@@ -16,7 +16,8 @@ pub use raftstore2::coprocessor::split_observer;
 pub use raftstore2::coprocessor::config::Config;
 pub use self::dispatcher::{CoprocessorHost, Registry};
 pub use raftstore2::coprocessor::error::{Error, Result};
-pub use self::region_info_accessor::{RegionInfo, RegionInfoAccessor, SeekRegionCallback};
+pub use tikv_misc::kv_region_info::RegionInfo;
+pub use raftstore2::coprocessor::region_info_accessor::{RegionInfoAccessor, SeekRegionCallback};
 pub use self::split_check::{
     get_region_approximate_keys, get_region_approximate_keys_cf, get_region_approximate_middle,
     get_region_approximate_size, get_region_approximate_size_cf, HalfCheckObserver,
